+++ EssentialFeed/EssentialFeediOSTests/Helpers/FeedViewControllerTests+LoaderSpy.swift
@@ -1,0 +1,69 @@
+//
+//  FeedViewControllerTests+LoaderSpy.swift
+//  EssentialFeediOSTests
+//
+//  Created by Arifin Firdaus on 17/07/21.
+//  Copyright © 2021 arifinfrds. All rights reserved.
+//
+
+import Foundation
+import EssentialFeed
+import EssentialFeediOS
+
+class LoaderSpy: FeedLoader, FeedImageDataLoader {
+    
+    // MARK: - FeedLoader
+    
+    private var feedRequests = [(FeedLoader.Result) -> Void]()
+    
+    var loadFeedCallCount: Int {
+        return feedRequests.count
+    }
+    
+    func load(completion: @escaping (FeedLoader.Result) -> Void) {
+        feedRequests.append(completion)
+    }
+    
+    func completeFeedLoading(with feed: [FeedImage] = [], at index: Int = 0) {
+        feedRequests[index](.success(feed))
+    }
+    
+    func completeFeedLoadingWithError(at index: Int = 0) {
+        let error = NSError(domain: "an error", code: 0)
+        feedRequests[index](.failure(error))
+    }
+    
+    // MARK: - FeedImageDataLoader
+    
+    private struct TaskSpy: FeedImageDataLoaderTask {
+        let cancelCallback: () -> Void
+        func cancel() {
+            cancelCallback()
+        }
+    }
+    
+    private var imageRequests = [(url: URL, completion: (FeedImageDataLoader.Result) -> Void)]()
+    
+    var loadedImageURLs: [URL] {
+        return imageRequests.map { $0.url }
+    }
+    private(set) var cancelledImageURLs = [URL]()
+    
+    func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
+        imageRequests.append((url, completion))
+        return TaskSpy { [weak self] in self?.cancelledImageURLs.append(url) }
+    }
+    
+    func cancelImageDataLoad(from url: URL) {
+        cancelledImageURLs.append(url)
+    }
+    
+    func completeImageLoading(with imageData: Data = Data(), at index: Int = 0) {
+        imageRequests[index].completion(.success(imageData))
+    }
+    
+    func completeImageLoadingWithError(at index: Int = 0) {
+        let error = NSError(domain: "an error", code: 0)
+        imageRequests[index].completion(.failure(error))
+    }
+}
